requirements.txt
Django==3.2.23
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
django-otp==1.2.2
django-filter==23.5
django-cors-headers==4.3.1
python-dotenv==1.0.1
google-generativeai==0.3.2
requests==2.31.0
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.34
celery==5.3.6
redis==5.0.1
django-celery-results==2.5.1
django-celery-beat==2.5.0
python-dateutil==2.8.2
pytz==2024.1 
dj-rest-auth